Update README.md
The ReadMe file is a very important part of the project which gives users instructions on how to set up and run this Project Management System in user-friendly manner.  

Some key points are the aim of the system, its key features, instructions to get started. The ReadMe file was created using Visual Studio Code	(version 1.95.3), a widely used and highly popular tool for creating ReadMe files.